feat: move WAF rule details into a per-rule edit modal
Replace the detail-heavy firewall table with a compact summary list
(ID, Description, Action + row actions) and edit each rule in a working-copy
modal that embeds the CEL condition builder. Retires the standalone
WafExpressionModal in favor of the in-modal builder.
